+++ src/main/java/model/entities/Category.java
@@ -2,7 +2,7 @@ package model.entities;
 
 public class Category {
 
-    private long category_id;
+    private int category_id;
     private String category_name;
 
     public Category() {}
@@ -11,16 +11,16 @@ public class Category {
         this.category_name = category_name;
     }
 
-    public Category(long category_id, String category_name) {
+    public Category(int category_id, String category_name) {
         this.category_id = category_id;
         this.category_name = category_name;
     }
 
-    public long getCategory_id() {
+    public int getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(long category_id) {
+    public void setCategory_id(int category_id) {
         this.category_id = category_id;
     }
