+++ src/main/java/model/entities/Question.java
@@ -2,8 +2,8 @@ package model.entities;
 
 public class Question {
 
-    private long question_id;
-    private long category_id;
+    private int question_id;
+    private int category_id;
     private String question_text;
 
     public Question() {}
@@ -12,30 +12,30 @@ public class Question {
         this.question_text = question_text;
     }
 
-    public Question(long category_id, String question_text) {
+    public Question(int category_id, String question_text) {
         this.category_id = category_id;
         this.question_text = question_text;
     }
 
-    public Question(long question_id, long category_id, String question_text) {
+    public Question(int question_id, int category_id, String question_text) {
         this.question_id = question_id;
         this.category_id = category_id;
         this.question_text = question_text;
     }
 
-    public long getQuestion_id() {
+    public int getQuestion_id() {
         return question_id;
     }
 
-    public void setQuestion_id(long question_id) {
+    public void setQuestion_id(int question_id) {
         this.question_id = question_id;
     }
 
-    public long getCategory_id() {
+    public int getCategory_id() {
         return category_id;
     }
 
-    public void setCategory_id(long category_id) {
+    public void setCategory_id(int category_id) {
         this.category_id = category_id;
     }
